+++ mod-erm-usage-server/src/test/java/org/folio/rest/impl/AggregatorSettingsIT.java
@@ -1,4 +1,4 @@
-package org.folio.mod_erm_usage_test;
+package org.folio.rest.impl;
 
 import static io.restassured.RestAssured.given;
 import static org.assertj.core.api.Assertions.assertThat;
